+++ recipe/common.php
@@ -73,7 +73,7 @@ set('use_atomic_symlink', function () {
 set('composer_action', 'install');
 set('composer_options', '{{composer_action}} --verbose --prefer-dist --no-progress --no-interaction --no-dev --optimize-autoloader');
 
-set('env_vars', ''); // Variable assignment before cmds (for example, SYMFONY_ENV={{set}})
+set('env', []); // Run command environment (for example, SYMFONY_ENV=prod)
 
 /**
  * Return current release path.